Disable non-rectangular selection from paste handling.
